fix js
