+++ react-components/src/__tests__/SyncChannel.test.js
@@ -3,16 +3,16 @@ import Enzyme from 'enzyme';
 
 import Adapter from 'enzyme-adapter-react-16'
 
-import { SelectChannelContainer } from '../../../react-components/src/components/SelectChannel';
+import { SyncChannelContainer } from '../components/SyncChannel';
 
 const setUp = (props = {}) => {
-  const component = Enzyme.shallow(<SelectChannelContainer {...props} />);
+  const component = Enzyme.shallow(<SyncChannelContainer {...props} />);
   return component;
 }
 
 Enzyme.configure({ adapter: new Adapter() })
 
-describe('SelectChannel-Component', () => {
+describe('SyncChannel-Component', () => {
   let component;
 
   beforeEach(() => {
